Simple refactor. Now protecting ROM from being written to. TETRIS runs

+++ include/timer.h
@@ -10,7 +10,7 @@ class Timer{
 		void process(Ticks ticks);
 	
 	private:
-		unsigned int clock();
+		uint16_t clock();
 		//Timer ticks
 		Ticks t;
 		//DIV ticks
